refactor relevant classes into  /util (#221)
* refactor JSONStandardParser.kt into /util

* refactor MyFragment.kt into /util

* refactor SetUtility.kt into /util

* refactor StringComparisons.kt into /util

* refactor ViewHolder.kt into /util

* fix missing StringComparisons imports
