+++ app/src/main/java/com/github/fribourgsdp/radio/MusixmatchLyricsGetter.kt
@@ -1,12 +1,13 @@
 package com.github.fribourgsdp.radio
 
 import android.util.Log
+import com.github.fribourgsdp.radio.util.JSONParser
+import com.github.fribourgsdp.radio.util.JSONStandardParser
 import okhttp3.*
 import org.json.JSONArray
 import java.io.IOException
 import java.util.*
 import java.util.concurrent.CompletableFuture
-import java.util.concurrent.Future
 
 const val API_KEY = "a3454edb65483e706c127deaa11df69d"
 const val BASE_URL = "http://api.musixmatch.com/ws/1.1/"
