+++ app/src/main/java/com/github/fribourgsdp/radio/PlayerGameHandler.kt
@@ -1,6 +1,8 @@
 package com.github.fribourgsdp.radio
 
 import com.github.fribourgsdp.radio.data.User
+import com.github.fribourgsdp.radio.util.NOT_THE_SAME
+import com.github.fribourgsdp.radio.util.StringComparisons
 import com.google.firebase.firestore.DocumentSnapshot
 
 class PlayerGameHandler(
