+++ app/src/main/java/com/github/fribourgsdp/radio/data/view/UserProfileActivity.kt
@@ -10,6 +10,7 @@ import android.widget.ImageView
 import android.widget.TextView
 import com.github.fribourgsdp.radio.*
 import com.github.fribourgsdp.radio.data.User
+import com.github.fribourgsdp.radio.util.MyFragment
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 import com.google.firebase.auth.FirebaseAuth
 import com.spotify.sdk.android.auth.AuthorizationClient
