+++ app/src/main/java/com/github/fribourgsdp/radio/util/MyFragment.kt
@@ -1,4 +1,4 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.util
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,6 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
+import com.github.fribourgsdp.radio.R
 
 abstract class MyFragment(private val layoutID : Int) : Fragment() {
     override fun onCreateView(inflater: LayoutInflater,
