+++ app/src/main/java/com/github/fribourgsdp/radio/util/StringComparisons.kt
@@ -1,4 +1,4 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.util
 
 import java.text.Normalizer
